function_app/requirements.txt
azure-functions
requests
azure-identity
azure-search-documents
azure-cosmos
azure-core